LSD-102 - EmailHandler moved to util
